Move ThemeTraits under one directory
