+++ web/modules/custom/server_general/src/ThemeTrait/CarouselTrait.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Drupal\server_general\ElementTrait;
+namespace Drupal\server_general\ThemeTrait;
 
 use Drupal\server_general\ElementLayoutTrait;
 use Drupal\server_general\ElementWrapTrait;
