+++ web/modules/custom/server_general/src/ThemeTrait/LineSeparatorTrait.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace Drupal\server_general;
+declare(strict_types=1);
+
+namespace Drupal\server_general\ThemeTrait;
 
 /**
  * Helper methods for a line separator.
