+++ web/modules/custom/server_general/src/ThemeTrait/TitleAndLabelsTrait.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace Drupal\server_general;
+declare(strict_types=1);
+
+namespace Drupal\server_general\ThemeTrait;
 
 /**
  * Helper method for building Title and labels of a content.
